LocalDB Common Managed Code and ARM Support (with LocalDB.exe)
