+++ src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClient/SNI/LocalDB.uap.cs
@@ -2,11 +2,18 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System;
+
 namespace Microsoft.Data.SqlClient.SNI
 {
     internal class LocalDB
     {
         internal static string GetLocalDBConnectionString(string localDbInstance)
+        {
+            throw new PlatformNotSupportedException(Strings.LocalDBNotSupported); // No Registry support on UAP
+        }
+
+        internal static string GetLocalDBDataSource(string fullServerName, out bool error)
         {
             throw new PlatformNotSupportedException(Strings.LocalDBNotSupported); // No Registry support on UAP
         }
