+++ src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClient/SNI/SNICommon.cs
@@ -119,14 +119,7 @@ namespace Microsoft.Data.SqlClient.SNI
         internal const int MultiSubnetFailoverWithInstanceSpecified = 48;
         internal const int MultiSubnetFailoverWithNonTcpProtocol = 49;
         internal const int MaxErrorValue = 50157;
-        internal const int LocalDBNoInstanceName = 51;
-        internal const int LocalDBNoInstallation = 52;
-        internal const int LocalDBInvalidConfig = 53;
-        internal const int LocalDBNoSqlUserInstanceDllPath = 54;
-        internal const int LocalDBInvalidSqlUserInstanceDllPath = 55;
-        internal const int LocalDBFailedToLoadDll = 56;
-        internal const int LocalDBBadRuntime = 57;
-
+   
         /// <summary>
         /// We only validate Server name in Certificate to match with "targetServerName".
         /// Certificate validation and chain trust validations are done by SSLStream class [System.Net.Security.SecureChannel.VerifyRemoteCertificate method]
